+++ components/merkledb/src/db.rs
@@ -616,6 +616,23 @@ impl Fork {
     pub fn working_patch(&self) -> &WorkingPatch {
         &self.working_patch
     }
+
+    /// Creates a fork based on the provided `patch` and `snapshot`.
+    ///
+    /// Note: using created fork to modify data already present in `patch` may lead
+    /// to an inconsistent database state. Hence, this method is useful only if you 
+    /// are sure that the fork and `patch` interacted with different indices.
+    pub fn from_patch(patch: Patch, snapshot: Box<dyn Snapshot>) -> Self {
+        let flushed = FlushedFork {
+            snapshot,
+            patch,
+        };
+
+        Self {
+            flushed,
+            working_patch: WorkingPatch::new(),
+        }
+    }
 }
 
 impl<'a> IndexAccess for &'a Fork {